in MmapSliceBufferedUpdateWrapper

+++ lib/segment/src/id_tracker/immutable_id_tracker.rs
@@ -398,7 +398,7 @@ impl IdTracker for ImmutableIdTracker {
         if has_version {
             self.internal_to_version.set(internal_id, version);
             self.internal_to_version_wrapper
-                .set(internal_id as usize, version);
+                .set(internal_id, version);
         }
 
         Ok(())
